=move punk classifiers out of classifer.py
